avances en streamlit, ignore de joblib
