+++ .gitattributes
@@ -1,0 +1,1 @@
+*.joblib filter=lfs diff=lfs merge=lfs -text
