+++ Streamlit/src/pages/requirements.txt
@@ -1,0 +1,107 @@
+holidays
+tensorflow 
+scikit-learn
+keras
+scikeras
+openmeteo_requests
+requests_cache
+retry_requests
+joblib
+xgboost
+lightgbm
+plotly
+altair==5.3.0
+asttokens==2.4.1
+attrs==23.2.0
+blinker==1.7.0
+cachetools==5.3.3
+certifi==2024.2.2
+charset-normalizer==3.3.2
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+colorama==0.4.6
+comm==0.2.2
+contourpy==1.2.1
+cycler==0.12.1
+debugpy==1.8.1
+decorator==5.1.1
+et-xmlfile==1.1.0
+exceptiongroup==1.2.0
+executing==2.0.1
+fiona==1.9.6
+fonttools==4.51.0
+geopandas==0.14.3
+gitdb==4.0.11
+GitPython==3.1.43
+google-api-core==2.18.0
+google-auth==2.29.0
+google-cloud==0.34.0
+google-cloud-bigquery==3.20.1
+google-cloud-core==2.4.1
+google-cloud-storage==2.16.0
+google-crc32c==1.5.0
+google-resumable-media==2.7.0
+googleapis-common-protos==1.63.0
+grpcio==1.62.1
+grpcio-status==1.62.1
+idna==3.7
+ipykernel==6.29.4
+ipython==8.23.0
+jedi==0.19.1
+Jinja2==3.1.3
+jsonschema==4.21.1
+jsonschema-specifications==2023.12.1
+jupyter_client==8.6.1
+jupyter_core==5.7.2
+kiwisolver==1.4.5
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.8.4
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+nest-asyncio==1.6.0
+numpy==1.26.4
+openpyxl==3.1.2
+packaging==24.0
+pandas==2.2.2
+parso==0.8.4
+pillow==10.3.0
+platformdirs==4.2.0
+prompt-toolkit==3.0.43
+proto-plus==1.23.0
+protobuf==4.25.3
+psutil==5.9.8
+pure-eval==0.2.2
+pyarrow==15.0.2
+pyasn1==0.6.0
+pyasn1_modules==0.4.0
+pydeck==0.8.1b0
+Pygments==2.17.2
+pyparsing==3.1.2
+pyproj==3.6.1
+python-dateutil==2.9.0.post0
+pytz==2024.1
+pywin32==306
+pyzmq==26.0.0
+referencing==0.34.0
+requests==2.31.0
+rich==13.7.1
+rpds-py==0.18.0
+rsa==4.9
+seaborn==0.13.2
+shapely==2.0.4
+six==1.16.0
+smmap==5.0.1
+stack-data==0.6.3
+streamlit==1.33.0
+tenacity==8.2.3
+toml==0.10.2
+toolz==0.12.1
+tornado==6.4
+traitlets==5.14.2
+typing_extensions==4.11.0
+tzdata==2024.1
+urllib3==2.2.1
+watchdog==4.0.0
+wcwidth==0.2.13
